refactor(api) move YamlParsingError to src
