+++ api/lib/domain/services/solution-service-qrocm-ind.js
@@ -2,10 +2,10 @@ import levenshtein from 'fast-levenshtein';
 import jsYaml from 'js-yaml';
 
 import { LEVENSHTEIN_DISTANCE_MAX_RATE } from '../../../src/shared/domain/constants.js';
+import { YamlParsingError } from '../../../src/shared/domain/errors.js';
 import { AnswerStatus } from '../../../src/shared/domain/models/AnswerStatus.js';
 import { _ } from '../../../src/shared/infrastructure/utils/lodash-utils.js';
 import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
-import { YamlParsingError } from '../../domain/errors.js';
 import { useLevenshteinRatio } from './services-utils.js';
 import { applyPreTreatments, applyTreatments } from './validation-treatments.js';
