+++ api/tests/unit/domain/services/solution-service-qrocm-dep_test.js
@@ -1,9 +1,9 @@
-import { YamlParsingError } from '../../../../lib/domain/errors.js';
 import {
   getCorrection,
   getCorrectionDetails,
   match,
 } from '../../../../lib/domain/services/solution-service-qrocm-dep.js';
+import { YamlParsingError } from '../../../../src/shared/domain/errors.js';
 import { AnswerStatus } from '../../../../src/shared/domain/models/AnswerStatus.js';
 import { catchErr, expect, sinon } from '../../../test-helper.js';
